% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccl_cycles}
\alias{ccl_cycles}
\title{Connected component border vertice paths.}
\usage{
ccl_cycles(m, reverse = FALSE)
}
\arguments{
\item{m}{A two dimensional numeric matrix.}

\item{reverse}{Begin at the minimal column major indexed vertex.}
}
\value{
Nested list of lists of arrays for all outlines of each
component such that \code{ccl_cycles(mat)[[label]][[id]]} yields
an array in the form of
\code{[ v_1, v_2, v_3, ... , v_1]}.

Each [[label]] index consists of at least 1 path and if the
component has holes there is an additional path for each hole.
}
\description{
Connected component grid graph vertice paths for
each eulerian cycle for each component.
}
\details{
There is a slight performance hit with `reverse` as this
forces a scan of the vertices for the minimal remaining valued
vertice and reorders the final results.
}
