% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccl_edges}
\alias{ccl_edges}
\title{Connected component border grid graph edges.}
\usage{
ccl_edges(m)
}
\arguments{
\item{m}{A two dimensional numeric matrix.}
}
\value{
List of arrays for all directed edges for each connected
component such that \code{ccl_edges(mat)[[label]]} yields an
array in the form of
$[ from_1, to_1, from_2, to_2, ... , from_n, to_n]$.

The first two sets of the returned array correspond to the
starting index (in column major order) of the component
within \code{m}.
}
\description{
Connected component directed edge vertex
connections.
}

