% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccl_outlines}
\alias{ccl_outlines}
\title{Connected component border cartesian paths.}
\usage{
ccl_outlines(m, min_first = TRUE)
}
\arguments{
\item{min_first}{Begin at the minimal column major indexed vertex.}

\item{mat}{A two dimensional numeric matrix.}
}
\value{
Nested list of lists of matrices for all outlines of each
input label's connected components.

Each [[label]] index consists of at least 1 path and if the
component has holes there is an additional path for each hole.
}
\description{
Connected component grid graph cartesian coordinate
paths for each eulerian cycle (outline) for each component.
}
\details{
There is a performance hit with `min_first`.
}
\seealso{
ccl_borders, ccl_labels, ccl_vertices, ccl_cycles, ccl_paths
}
