% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccl_paths}
\alias{ccl_paths}
\title{Connected component border cartesian paths.}
\usage{
ccl_paths(m, reverse = FALSE)
}
\arguments{
\item{m}{A two dimensional numeric matrix.}

\item{reverse}{Begin at the minimal column major indexed vertex.}
}
\value{
Nested list of lists of arrays for all outlines of each
component such that \code{ccl_paths(mat)[[label]][[id]]}
yields an array in the form of
\code{[ v_1, y_1, x_1, v_2, y_2, x_2, ... , v_1, y_1, x_1]}.

Each [[label]] index consists of at least 1 path and if the
component has holes there is an addition  path for each hole.
}
\description{
Connected component grid graph cartesian coordinate
paths for each eulerian cycle (outline) for each component.
}
\details{
There is a performance hit with `reverse`.
}
\seealso{
ccl_cycles
}

