% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccloutline.R
\docType{package}
\name{ccloutline}
\alias{ccloutline}
\alias{ccloutline-package}
\title{Connected component labelling and eulerian cycle outlines.}
\description{
Given a 2 dimensional matrix generate the cartesian
coordinates to plot a grid graph path outlining the outer and inner
component edges.
}
\author{
Thell Fowler \email{thell@tbfowler.name}
}
\keyword{border}
\keyword{ccl}
\keyword{outline}
\keyword{package}

