% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccloutline.R
\docType{data}
\name{wiki_sample_mat}
\alias{wiki_sample_mat}
\title{Connected component extraction.}
\format{
A matrix 9 rows and 17 variables:
}
\source{
\url{https://en.wikipedia.org/wiki/Connected-component_labeling#Graphical_example_of_two-pass_algorithm}
}
\usage{
wiki_sample_mat
}
\description{
A matrix of binary values representing components.
}
\keyword{datasets}
